Extracted common base class of VstFxEditorView and VstiEditorView
Fix code duplication problems
